Task-4/requirements.txt
streamlit
faiss-cpu
beautifulsoup4
sentence-transformers
google-generativeai
numpy
requests
